Allow HAIstings to pick up where it left off
This allows HAIstings to continue the conversation, generating more
relevant reports, if there is something to pick up from.

Problematically, we can't use the streaming responses because langgraph
prints randomly the messages for some reason... Even if it should just
append. So, in order to keep state AND print only relevant bits, I moved
the printing logic to the nodes.

Signed-off-by: Juan Antonio Osorio <ozz@stacklok.com>